Harmonized the handling of symmetry in potentials

+++ src/fortran_wrapper.cpp
@@ -100,7 +100,7 @@ private:
 };
 
 /// convert FORTRAN string to C++ string
-static std::string stdstr(char* fortranString, int len)
+static std::string stdstr(char* fortranString, long len)
 {
     std::string str(len, '\0');
     memcpy(&(str[0]), fortranString, len);
@@ -117,7 +117,10 @@ std::vector<potential::PtrPotential> potentials;
 /// potential specified by parameters provided in an INI file.
 /// INPUT:  inifilename - a string with the name of INI file with all parameters.
 /// OUTPUT: c_obj - the placeholder for storing the pointer to the C++ potential object.
-extern "C" void agama_initfromfile_(void* c_obj, char* inifilename, int, int len)
+/// EXAMPLE IN FORTRAN:
+///     char*8 C_OBJ
+///     call agama_initfromfile(C_OBJ, "file.ini")  ! now C_OBJ contains a valid instance of potential
+extern "C" void agama_initfromfile_(void* c_obj, char* inifilename, long /*len(c_obj)=8*/, long len)
 {
     potentials.push_back(potential::createPotential(stdstr(inifilename, len)));
     const potential::BasePotential* ptr = potentials.back().get();
@@ -129,7 +132,7 @@ extern "C" void agama_initfromfile_(void* c_obj, char* inifilename, int, int len
 /// INPUT:  params - a string with parameters of the potential expansion
 /// (e.g., "type=Dehnen p=0.8 q=0.5 mass=10 scaleRadius=5").
 /// OUTPUT: c_obj - the placeholder for storing the pointer to the C++ potential object.
-extern "C" void agama_initfromparam_(void* c_obj, char* params, int, int len)
+extern "C" void agama_initfromparam_(void* c_obj, char* params, long, long len)
 {
     potentials.push_back(potential::createPotential(utils::KeyValueMap(stdstr(params, len))));
     const potential::BasePotential* ptr = potentials.back().get();
@@ -142,7 +145,7 @@ extern "C" void agama_initfromparam_(void* c_obj, char* params, int, int len)
 /// (e.g., "type=Multipole rmin=0.01 rmax=100 lmax=8 mmax=6 symmetry=Triaxial").
 /// INPUT:  pot  - the FORTRAN routine that computes the potential and its derivative;
 /// OUTPUT: c_obj - the placeholder for storing the pointer to the C++ potential object.
-extern "C" void agama_initfrompot_(void* c_obj, char* params, potentialfnc pot, int, int len)
+extern "C" void agama_initfrompot_(void* c_obj, char* params, potentialfnc pot, long, long len)
 {
     utils::KeyValueMap param(stdstr(params, len));
     PotentialWrapper wrapper(pot, potential::getSymmetryTypeByName(param.getString("Symmetry")));
@@ -152,7 +155,7 @@ extern "C" void agama_initfrompot_(void* c_obj, char* params, potentialfnc pot, 
 }
 
 /// same as above, but for a user-defined density function
-extern "C" void agama_initfromdens_(void* c_obj, char* paramstr, densityfnc pot, int, int len)
+extern "C" void agama_initfromdens_(void* c_obj, char* paramstr, densityfnc pot, long, long len)
 {
     utils::KeyValueMap param(stdstr(paramstr, len));
     DensityWrapper wrapper(pot, potential::getSymmetryTypeByName(param.getString("Symmetry")));
@@ -165,7 +168,7 @@ extern "C" void agama_initfromdens_(void* c_obj, char* paramstr, densityfnc pot,
 /// INPUT:  c_obj  is the placeholder for the pointer to a previously created potential.
 /// INPUT:  X[3]   is the array of coordinates (x,y,z).
 /// RETURN: the potential at the point Phi(x,y,z).
-extern "C" double agama_potential_(void* c_obj, double* X, int)
+extern "C" double agama_potential_(void* c_obj, double* X, long)
 {
     const potential::BasePotential* pot;
     memcpy(&pot, c_obj, sizeof(void*));
@@ -177,7 +180,7 @@ extern "C" double agama_potential_(void* c_obj, double* X, int)
 /// INPUT:  X[3]   is the array of coordinates (x,y,z).
 /// OUTPUT: FORCE[3] will contain the force -dPhi/dx, etc.
 /// RETURN: the potential at the point Phi(x,y,z).
-extern "C" double agama_potforce_(void* c_obj, double* X, double* FORCE, int)
+extern "C" double agama_potforce_(void* c_obj, double* X, double* FORCE, long)
 {
     const potential::BasePotential* pot;
     memcpy(&pot, c_obj, sizeof(void*));
@@ -197,7 +200,7 @@ extern "C" double agama_potforce_(void* c_obj, double* X, double* FORCE, int)
 /// OUTPUT: DERIV[6] will contain the force derivatives in the following order:
 /// dFx/dx, dFy/dy, dFz/dz, dFx/dy, dFx/dz, dFy/dz.
 /// RETURN: the potential at the point Phi(x,y,z).
-extern "C" double agama_potforcederiv_(void* c_obj, double* X, double* FORCE, double* DERIV, int)
+extern "C" double agama_potforcederiv_(void* c_obj, double* X, double* FORCE, double* DERIV, long)
 {
     const potential::BasePotential* pot;
     memcpy(&pot, c_obj, sizeof(void*));
@@ -221,7 +224,7 @@ extern "C" double agama_potforcederiv_(void* c_obj, double* X, double* FORCE, do
 // INPUT:  c_obj  is the placeholder for the pointer to a previously created potential.
 // INPUT:  X[3]   is the array of coordinates (x,y,z).
 // RETURN: the value of density.
-extern "C" double agama_density_(void* c_obj, double* X, int)
+extern "C" double agama_density_(void* c_obj, double* X, long)
 {
     const potential::BasePotential* pot;
     memcpy(&pot, c_obj, sizeof(void*));
